host/pyproject.toml
[project]
name = "software"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "dearpygui>=2.0.0",
    "docopt>=0.6.2",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "pygame>=2.6.1",
    "pymvg>=2.1.0",
    "pyqt6>=6.9.0",
    "pyserial-asyncio>=0.6",
    "scikit-learn>=1.7.0",
    "vispy>=0.14.3",
]
